header management API consistency
extract common code for get/set/unset/getHeader to request-base.js
extract is-object to standalone commonjs module
fixes #873
part of #830
