+++ lib/node/index.js
@@ -24,6 +24,7 @@ var zlib = require('zlib');
 var util = require('util');
 var pkg = require('../../package.json');
 var requestBase = require('../request-base');
+var isObject = require('../is-object');
 
 /**
  * Expose the request function.
@@ -73,18 +74,6 @@ exports.protocols = {
 };
 
 /**
- * Check if `obj` is an object.
- *
- * @param {Object} obj
- * @return {Boolean}
- * @api private
- */
-
-function isObject(obj) {
-  return null != obj && 'object' == typeof obj;
-}
-
-/**
  * Default serialization map.
  *
  *     superagent.serialize['application/xml'] = function(obj){
@@ -110,6 +99,22 @@ exports.serialize = {
 exports.parse = require('./parsers');
 
 /**
+ * Initialize internal header tracking properties on a request instance.
+ *
+ * @param {Object} req the instance
+ * @api private
+ */
+function _initHeaders(req) {
+  var ua = 'node-superagent/' + pkg.version;
+  req._header = { // coerces header names to lowercase
+    'user-agent': ua
+  };
+  req.header = { // preserves header name case
+    'User-Agent': ua
+  };
+}
+
+/**
  * Initialize a new `Request` with the given `method` and `url`.
  *
  * @param {String} method
@@ -125,9 +130,7 @@ function Request(method, url) {
   this._formData = null;
   this.method = method;
   this.url = url;
-  this.header = {
-    'user-agent': 'node-superagent/' + pkg.version
-  };
+  _initHeaders(this);
   this.writable = true;
   this._redirects = 0;
   this.redirects(5);
@@ -249,75 +252,6 @@ Request.prototype.agent = function(agent){
   if (!arguments.length) return this._agent;
   this._agent = agent;
   return this;
-};
-
-/**
- * Set header `field` to `val`, or multiple fields with one object.
- * Case-insensitive.
- *
- * Examples:
- *
- *      req.get('/')
- *        .set('Accept', 'application/json')
- *        .set('X-API-Key', 'foobar')
- *        .end(callback);
- *
- *      req.get('/')
- *        .set({ Accept: 'application/json', 'X-API-Key': 'foobar' })
- *        .end(callback);
- *
- * @param {String|Object} field
- * @param {String} val
- * @return {Request} for chaining
- * @api public
- */
-
-Request.prototype.set = function(field, val){
-  if (isObject(field)) {
-    for (var key in field) {
-      this.set(key, field[key]);
-    }
-    return this;
-  }
-
-  debug('set %s "%s"', field, val);
-  this.header[field.toLowerCase()] = val;
-  return this;
-};
-
-/**
- * Remove header `field`.
- * Case-insensitive.
- *
- * Example:
- *
- *      req.get('/')
- *        .unset('User-Agent')
- *        .end(callback);
- *
- * @param {String} field
- * @return {Request} for chaining
- * @api public
- */
-
-Request.prototype.unset = function(field){
-  debug('unset %s', field);
-
-  delete this.header[field.toLowerCase()];
-  return this;
-};
-
-/**
- * Get request header `field`.
- * Case-insensitive.
- *
- * @param {String} field
- * @return {String}
- * @api public
- */
-
-Request.prototype.get = function(field){
-  return this.header[field.toLowerCase()];
 };
 
 /**
@@ -444,7 +378,7 @@ Request.prototype.query = function(val){
 
 Request.prototype.send = function(data){
   var obj = isObject(data);
-  var type = this.get('Content-Type');
+  var type = this._header['content-type'];
   // merge
   if (obj && isObject(this._data)) {
     for (var key in data) {
@@ -454,7 +388,7 @@ Request.prototype.send = function(data){
   } else if ('string' == typeof data) {
     // default to x-www-form-urlencoded
     if (!type) this.type('form');
-    type = this.get('Content-Type');
+    type = this._header['content-type'];
 
     // concat &
     if ('application/x-www-form-urlencoded' == type) {
@@ -620,11 +554,7 @@ Request.prototype.redirect = function(res){
   delete this._formData;
 
   // remove all add header except User-Agent
-  for (var key in this.header) {
-    if (key !== 'User-Agent') {
-      delete this.header[key]
-    }
-  }
+  _initHeaders(this)
 
   // redirect
   this.url = url;
